Resolve build errors with JS compilation
