--- index.ts
@@ -1,14 +1,0 @@
-import { CamerasTranslator, DoodadsTranslator, ImportsTranslator, InfoTranslator, ObjectsTranslator, RegionsTranslator, SoundsTranslator, StringsTranslator, TerrainTranslator, UnitsTranslator } from './lib/translators';
-
-export {
-    CamerasTranslator,
-    DoodadsTranslator,
-    ImportsTranslator,
-    InfoTranslator,
-    ObjectsTranslator,
-    RegionsTranslator,
-    SoundsTranslator,
-    StringsTranslator,
-    TerrainTranslator,
-    UnitsTranslator
-};
