+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "wc3maptranslator",
-    "version": "4.0.1",
+    "version": "4.0.2",
     "description": "Module to translate between `war3map` and `json` formats for WarCraft III .w3x maps",
     "keywords": [
         "wc3",
@@ -13,11 +13,12 @@
         "json",
         "w3x"
     ],
-    "main": "index.js",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "files": ["dist/**"],
     "scripts": {
         "build": "tsc --build",
         "lint": "tslint --project .",
-        "postinstall": "npm run build",
         "test": "npm run build && mocha",
         "test-nyc": "npm run build && nyc mocha",
         "test-travis": "npm run build && ./node_modules/istanbul/lib/cli.js cover ./node_modules/mocha/bin/_mocha --include-all-sources true --root ./lib -- -R spec ./test/*.js"
@@ -45,6 +46,7 @@
     },
     "homepage": "https://github.com/ChiefOfGxBxL/WC3MapTranslator#readme",
     "dependencies": {
+        "ieee754": "^1.2.1",
         "intn": "^1.0.0",
         "round-to": "^5.0.0"
     },
@@ -57,7 +59,6 @@
         "coveralls": "^3.1.0",
         "diff-buf": "^1.0.1",
         "fs-extra": "^9.1.0",
-        "ieee754": "^1.2.1",
         "istanbul": "^0.4.5",
         "mocha": "^8.3.0",
         "nyc": "^15.1.0",
