+++ src/AngleConverter.ts
