+++ src/HexBuffer.ts
