+++ src/index.ts
@@ -1,0 +1,25 @@
+import {
+    CamerasTranslator,
+    DoodadsTranslator,
+    ImportsTranslator,
+    InfoTranslator, 
+    ObjectsTranslator,
+    RegionsTranslator,
+    SoundsTranslator,
+    StringsTranslator,
+    TerrainTranslator,
+    UnitsTranslator
+} from './translators';
+
+export {
+    CamerasTranslator,
+    DoodadsTranslator,
+    ImportsTranslator,
+    InfoTranslator,
+    ObjectsTranslator,
+    RegionsTranslator,
+    SoundsTranslator,
+    StringsTranslator,
+    TerrainTranslator,
+    UnitsTranslator
+};
