+++ src/translators/CamerasTranslator.ts
