+++ src/translators/DoodadsTranslator.ts
