+++ src/translators/InfoTranslator.ts
