+++ src/translators/RegionsTranslator.ts
