+++ src/translators/SoundsTranslator.ts
