+++ src/translators/StringsTranslator.ts
