+++ src/translators/TerrainTranslator.ts
