+++ test/TranslatorReversion.ts
@@ -3,7 +3,7 @@ import * as diff from 'diff-buf';
 import * as fs from 'fs-extra';
 import * as Path from 'path';
 
-import * as Translator from '../index';
+import * as Translator from '../src';
 
 const ObjectType = Translator.ObjectsTranslator.ObjectType;
 const war3mapDir = Path.resolve('test/data');
